pyproject.toml
[build-system]
requires = ["hatchling", "setuptools>=65.0", "wheel"]
build-backend = "hatchling.build"

[project]
name = "pymonitor"
version = "0.1.0"
description = "A Python monitoring tool for tracking function execution"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "LiveProbes" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "bytecode>=0.16.2",
    "chlorophyll",
    "debugpy>=1.8.14",
    "fastapi>=0.115.12",
    "fastparquet>=2024.11.0",
    "flask>=2.0.0",
    "flask-cors>=4.0.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "networkx>=3.5",
    "pandas>=2.3.1",
    "pillow>=11.2.1",
    "pipreqs>=0.5.0",
    "pyarrow>=20.0.0",
    "pygame>=2.6.1",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.34.2",
    "watchdog>=6.0.0",
]

[project.scripts]
web-pymonitor = "monitoringpy.interface.web.explorer:main"
game-explorer = "monitoringpy.interface.gameexplorer.gameexplorer:main"
game-explorer2 = "monitoringpy.interface.gameexplorer.gameexplorer2:main"

[tool.hatch.build.targets.wheel]
packages = ["src/monitoringpy"]
include = [
    "src/monitoringpy/codediff/download_gumtree.py",
    "src/monitoringpy/codediff/gumtree_utils.py",
] 

[tool.uv.sources]
chlorophyll = { git = "https://github.com/jbdoderlein/chlorophyll" }

[tool.ruff.lint]
extend-select = [
    "F",        # Pyflakes rules
    "W",        # PyCodeStyle warnings
    "E",        # PyCodeStyle errors
    "I",        # Sort imports properly
    "UP",       # Warn if certain things can changed due to newer Python versions
    "C4",       # Catch incorrect use of comprehensions, dict, list, etc
    "FA",       # Enforce from __future__ import annotations
    "ISC",      # Good use of string concatenation
    "ICN",      # Use common import conventions
    "RET",      # Good return practices
    "SIM",      # Common simplification rules
    "TID",      # Some good import practices
    "TC",       # Enforce importing certain types in a TYPE_CHECKING block
    "TD",       # Be diligent with TODO comments
]
ignore = [
    "COM812",  # missing-trailing-comma
    "CPY001",  # Missing copyright notice at top of file
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "D203",    # blank line required before class docstring
    "D211",    # no-blank-line-before-class
    "D213",    # multi-line-summary-second-line
    "EM101",   # raw-string-in-exception
    "FIX002",  # line-contains-todo
    "E501",    # Line too long
    "TD003",   # Missing issue link on todo
]
